src/test/java/seedu/address/logic/commands/BusCommandTest.java
//@@author jshoung
package seedu.address.logic.commands;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertTrue;
import static seedu.address.logic.commands.BusCommand.SHOWING_BUS_MESSAGE;

import org.junit.Rule;
import org.junit.Test;

import seedu.address.commons.events.ui.ShowBusRequestEvent;
import seedu.address.ui.testutil.EventsCollectorRule;

public class BusCommandTest {
    @Rule
    public final EventsCollectorRule eventsCollectorRule = new EventsCollectorRule();

    @Test
    public void execute_bus_success() {
        CommandResult result = new BusCommand().execute();
        assertEquals(SHOWING_BUS_MESSAGE, result.feedbackToUser);
        assertTrue(eventsCollectorRule.eventsCollector.getMostRecent() instanceof ShowBusRequestEvent);
        assertTrue(eventsCollectorRule.eventsCollector.getSize() == 1);
    }
}
